# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a DDL Guard server API.",
        "title": "DDL Guard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/final-grades": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取用户所有期末成绩记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FinalGrade"
                ],
                "summary": "获取期末成绩列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.FinalGradeListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的期末成绩记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FinalGrade"
                ],
                "summary": "创建期末成绩",
                "parameters": [
                    {
                        "description": "Create Final Grade Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.CreateFinalGradeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.CreateFinalGradeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/final-grades/{uuid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取单个期末成绩记录的详细信息，包含关联的小测和作业成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FinalGrade"
                ],
                "summary": "获取期末成绩详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Final Grade UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.FinalGradeDetailResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新期末成绩记录的信息，自动重算最终成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FinalGrade"
                ],
                "summary": "更新期末成绩",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Final Grade UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Final Grade Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UpdateFinalGradeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.UpdateFinalGradeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除期末成绩记录及其关联的小测和作业成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FinalGrade"
                ],
                "summary": "删除期末成绩",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Final Grade UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "schema.CreateFinalGradeReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "attendance_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "classroom_bonus_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "exam_ratio": {
                    "type": "integer",
                    "example": 40
                },
                "homework_ratio": {
                    "type": "integer",
                    "example": 20
                },
                "name": {
                    "type": "string",
                    "example": "2024春季期末成绩"
                },
                "quiz_ratio": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "schema.CreateFinalGradeResp": {
            "type": "object",
            "properties": {
                "attendance_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "attendance_score": {
                    "type": "number",
                    "example": 95
                },
                "classroom_bonus_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "classroom_bonus_score": {
                    "type": "number",
                    "example": 90
                },
                "exam_ratio": {
                    "type": "integer",
                    "example": 40
                },
                "exam_score": {
                    "type": "number",
                    "example": 85.5
                },
                "final_score": {
                    "type": "number",
                    "example": 87
                },
                "homework_ratio": {
                    "type": "integer",
                    "example": 20
                },
                "name": {
                    "type": "string",
                    "example": "2024春季期末成绩"
                },
                "quiz_ratio": {
                    "type": "integer",
                    "example": 20
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
...
        "schema.FinalGradeDetailResp": {
            "type": "object",
            "properties": {
                "attendance_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "attendance_score": {
                    "type": "number",
                    "example": 95
                },
                "classroom_bonus_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "classroom_bonus_score": {
                    "type": "number",
                    "example": 90
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-20T10:00:00+08:00"
                },
                "exam_ratio": {
                    "type": "integer",
                    "example": 40
                },
                "exam_score": {
                    "type": "number",
                    "example": 85.5
                },
                "final_score": {
                    "type": "number",
                    "example": 87
                },
                "homework_ratio": {
                    "type": "integer",
                    "example": 20
                },
                "homework_scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.HomeworkScoreItem"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "2024春季期末成绩"
                },
                "quiz_ratio": {
                    "type": "integer",
                    "example": 20
                },
                "quiz_scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.QuizScoreItem"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-04-20T10:00:00+08:00"
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
        "schema.FinalGradeListItem": {
            "type": "object",
            "properties": {
                "attendance_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "attendance_score": {
                    "type": "number",
                    "example": 95
                },
                "classroom_bonus_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "classroom_bonus_score": {
                    "type": "number",
                    "example": 90
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-20T10:00:00+08:00"
                },
                "exam_ratio": {
                    "type": "integer",
                    "example": 40
                },
                "exam_score": {
                    "type": "number",
                    "example": 85.5
                },
                "final_score": {
                    "type": "number",
                    "example": 87
                },
                "homework_ratio": {
                    "type": "integer",
                    "example": 20
                },
                "name": {
                    "type": "string",
                    "example": "2024春季期末成绩"
                },
                "quiz_ratio": {
                    "type": "integer",
                    "example": 20
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
        "schema.FinalGradeListResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.FinalGradeListItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "schema.HomeworkScoreItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "作业1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
...
        "schema.QuizScoreItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "小测1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
...
        "schema.UpdateFinalGradeReq": {
            "type": "object",
            "properties": {
                "attendance_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "attendance_score": {
                    "type": "number",
                    "example": 95
                },
                "classroom_bonus_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "classroom_bonus_score": {
                    "type": "number",
                    "example": 90
                },
                "exam_ratio": {
                    "type": "integer",
                    "example": 40
                },
                "exam_score": {
                    "type": "number",
                    "example": 85.5
                },
                "homework_ratio": {
                    "type": "integer",
                    "example": 20
                },
                "name": {
                    "type": "string",
                    "example": "2024春季期末成绩"
                },
                "quiz_ratio": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "schema.UpdateFinalGradeResp": {
            "type": "object",
            "properties": {
                "attendance_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "attendance_score": {
                    "type": "number",
                    "example": 95
                },
                "classroom_bonus_ratio": {
                    "type": "integer",
                    "example": 10
                },
                "classroom_bonus_score": {
                    "type": "number",
                    "example": 90
                },
                "exam_ratio": {
                    "type": "integer",
                    "example": 40
                },
                "exam_score": {
                    "type": "number",
                    "example": 85.5
                },
                "final_score": {
                    "type": "number",
                    "example": 87
                },
                "homework_ratio": {
                    "type": "integer",
                    "example": 20
                },
                "name": {
                    "type": "string",
                    "example": "2024春季期末成绩"
                },
                "quiz_ratio": {
                    "type": "integer",
                    "example": 20
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}